Issue #SC-1727 chore: Refactor dp-core package names
