+++ data-pipeline-flink/de-normalization/src/main/scala/org/sunbird/dp/functions/LocationDenormFunction.scala
@@ -4,7 +4,7 @@ package org.sunbird.dp.functions
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.slf4j.LoggerFactory
 import org.sunbird.dp.task.DenormalizationConfig
-import org.sunbird.dp.core._
+import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 import org.sunbird.dp.domain.Event
 import org.apache.flink.streaming.api.functions.ProcessFunction
