+++ data-pipeline-flink/de-normalization/src/main/scala/org/sunbird/dp/task/DenormalizationConfig.scala
@@ -1,9 +1,8 @@
 package org.sunbird.dp.task
 
-import java.util
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.TypeExtractor
-import org.sunbird.dp.core.BaseJobConfig
+import org.sunbird.dp.core.job.BaseJobConfig
 import org.apache.flink.streaming.api.scala.OutputTag
 import com.typesafe.config.Config
 import org.sunbird.dp.domain.Event
