+++ data-pipeline-flink/de-normalization/src/test/scala/org/sunbird/dp/spec/DenormalizationStreamTaskTestSpec.scala
@@ -11,7 +11,6 @@ import org.apache.flink.streaming.api.functions.source.SourceFunction.SourceCont
 import org.apache.flink.test.util.MiniClusterWithClientResource
 import org.mockito.Mockito
 import org.mockito.Mockito._
-import org.sunbird.dp.core.FlinkKafkaConnector
 import org.sunbird.dp.domain.Event
 import org.sunbird.dp.fixture.EventFixture
 import org.sunbird.dp.task.DenormalizationConfig
@@ -20,7 +19,8 @@ import com.google.gson.Gson
 import com.typesafe.config.{Config, ConfigFactory}
 import org.sunbird.dp.{BaseMetricsReporter, BaseTestSpec}
 import redis.embedded.RedisServer
-import org.sunbird.dp.cache.RedisConnect
+import org.sunbird.dp.core.cache.RedisConnect
+import org.sunbird.dp.core.job.FlinkKafkaConnector
 
 class DenormalizationStreamTaskTestSpec extends BaseTestSpec {
