+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/domain/EventsPath.scala
@@ -1,4 +1,4 @@
-package org.sunbird.dp.domain
+package org.sunbird.dp.core.domain
 
 object EventsPath {
   val METADATA_PATH = "metadata"
