+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/job/FlinkKafkaConnector.scala
@@ -1,14 +1,13 @@
-package org.sunbird.dp.core
+package org.sunbird.dp.core.job
 
 import java.util
 
-import org.apache.flink.streaming.api.functions.source.SourceFunction
-import org.apache.flink.streaming.connectors.kafka.FlinkKafkaConsumer
-import org.sunbird.dp.serde._
 import org.apache.flink.streaming.api.functions.sink.SinkFunction
-import org.apache.flink.streaming.connectors.kafka.FlinkKafkaProducer
+import org.apache.flink.streaming.api.functions.source.SourceFunction
+import org.apache.flink.streaming.connectors.kafka.{FlinkKafkaConsumer, FlinkKafkaProducer}
 import org.apache.flink.streaming.connectors.kafka.FlinkKafkaProducer.Semantic
-import org.sunbird.dp.domain.Events
+import org.sunbird.dp.core.domain.Events
+import org.sunbird.dp.core.serde._
 
 class FlinkKafkaConnector(config: BaseJobConfig) extends Serializable {
