+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/reader/ParentMap.scala
@@ -1,4 +1,4 @@
-package org.sunbird.dp.reader
+package org.sunbird.dp.core.reader
 
 
 import java.util
