+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/reader/Telemetry.scala
@@ -1,4 +1,4 @@
-package org.sunbird.dp.reader
+package org.sunbird.dp.core.reader
 
 import java.io.Serializable
 import java.text.SimpleDateFormat
