+++ data-pipeline-flink/dp-core/src/main/scala/org/sunbird/dp/core/util/FlinkUtil.scala
@@ -1,9 +1,9 @@
-package org.sunbird.dp.util
+package org.sunbird.dp.core.util
 
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
-import org.sunbird.dp.core.BaseJobConfig
 import org.apache.flink.streaming.api.TimeCharacteristic
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
+import org.sunbird.dp.core.job.BaseJobConfig
 
 object FlinkUtil {
