+++ data-pipeline-flink/dp-core/src/test/scala/org/sunbird/spec/BaseProcessTestConfig.scala
@@ -6,7 +6,7 @@ import com.typesafe.config.Config
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.TypeExtractor
 import org.apache.flink.streaming.api.scala.OutputTag
-import org.sunbird.dp.core.BaseJobConfig
+import org.sunbird.dp.core.job.BaseJobConfig
 
 class BaseProcessTestConfig(override val config: Config) extends BaseJobConfig(config, "Test-job") {
   private val serialVersionUID = -2349318979085017498L
