+++ data-pipeline-flink/dp-core/src/test/scala/org/sunbird/spec/TestMapStreamFunc.scala
@@ -4,7 +4,7 @@ import java.util
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.functions.ProcessFunction
-import org.sunbird.dp.core.{BaseProcessFunction, Metrics}
+import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 
 
 class TestMapStreamFunc(config: BaseProcessTestConfig)(implicit val stringTypeInfo: TypeInformation[String])
