+++ data-pipeline-flink/dp-core/src/test/scala/org/sunbird/spec/TestStringStreamFunc.scala
@@ -2,8 +2,7 @@ package org.sunbird.spec
 
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.functions.ProcessFunction
-import org.sunbird.dp.core.{BaseProcessFunction, Metrics}
-
+import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 
 class TestStringStreamFunc(config: BaseProcessTestConfig)(implicit val stringTypeInfo: TypeInformation[String])
   extends BaseProcessFunction[String, String](config) {
