+++ data-pipeline-flink/druid-events-validator/src/main/scala/org/sunbird/dp/domain/Event.scala
@@ -2,6 +2,8 @@ package org.sunbird.dp.domain
 
 import java.util
 
+import org.sunbird.dp.core.domain.Events
+
 class Event(eventMap: util.Map[String, Any]) extends Events(eventMap) {
 
   private val jobName = "DruidValidator"
