+++ data-pipeline-flink/pipeline-preprocessor/src/main/scala/org/sunbird/dp/functions/ShareEventsFlattenerFunction.scala
@@ -7,7 +7,7 @@ import com.google.gson.Gson
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.streaming.api.functions.ProcessFunction
 import org.slf4j.LoggerFactory
-import org.sunbird.dp.core.{BaseProcessFunction, Metrics}
+import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 import org.sunbird.dp.domain.{Actor, EData, Event, Object, Rollup, ShareEvent}
 import org.sunbird.dp.task.PipelinePreprocessorConfig
