+++ data-pipeline-flink/pipeline-preprocessor/src/test/scala/org/sunbird/dp/spec/PipelineProcessorStreamTaskTestSpec.scala
@@ -14,9 +14,9 @@ import org.apache.flink.test.util.MiniClusterWithClientResource
 import org.mockito.Mockito
 import org.mockito.Mockito.when
 import org.sunbird.dp.{BaseMetricsReporter, BaseTestSpec}
-import org.sunbird.dp.core.FlinkKafkaConnector
 import org.sunbird.dp.domain.Event
 import org.sunbird.dp.fixture.EventFixtures
+import org.sunbird.dp.core.job.FlinkKafkaConnector
 import org.sunbird.dp.task.{PipelinePreprocessorConfig, PipelinePreprocessorStreamTask}
 import redis.embedded.RedisServer
