+++ data-pipeline-flink/telemetry-extractor/src/main/scala/org/sunbird/dp/functions/DeduplicationFunction.scala
@@ -6,9 +6,9 @@ import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.functions.ProcessFunction
 import org.slf4j.LoggerFactory
-import org.sunbird.dp.cache.{DedupEngine, RedisConnect}
+import org.sunbird.dp.core.cache.{DedupEngine, RedisConnect}
+import org.sunbird.dp.core.job.{BaseProcessFunction, Metrics}
 import org.sunbird.dp.task.TelemetryExtractorConfig
-import org.sunbird.dp.core.{BaseDeduplication, BaseProcessFunction, Metrics}
 
 class DeduplicationFunction(config: TelemetryExtractorConfig, @transient var dedupEngine: DedupEngine = null)
                            (implicit val mapTypeInfo: TypeInformation[util.Map[String, AnyRef]])
